+++ Controller/Payment/OpenOrder.php
@@ -8,7 +8,6 @@ use Magento\Framework\App\ResponseInterface;
 use Magento\Framework\Controller\Result\JsonFactory;
 use Nuvei\Checkout\Model\AbstractRequest;
 use Nuvei\Checkout\Model\Config as ModuleConfig;
-//use Nuvei\Checkout\Model\Payment;
 use Nuvei\Checkout\Model\Request\Factory as RequestFactory;
 
 /**
@@ -74,6 +73,7 @@ class OpenOrder extends Action
         $result = $this->jsonResultFactory->create()
             ->setHttpResponseCode(\Magento\Framework\Webapi\Response::HTTP_OK);
         
+        // if plugin is not active
         if (!$this->moduleConfig->getConfigValue('active')) {
             $this->readerWriter->createLog(
                 'OpenOrder error - '
@@ -81,9 +81,7 @@ class OpenOrder extends Action
             );
             
             return $result->setData(
-                [
-                'error_message' => __('OpenOrder error - Nuvei checkout module is not active at the moment!')
-                ]
+                ['error_message' => __('OpenOrder error - Nuvei checkout module is not active at the moment!')]
             );
         }
         
@@ -134,13 +132,11 @@ class OpenOrder extends Action
         }
         
         // success
-        return $result->setData(
-            [
+        return $result->setData([
             "error"         => 0,
             "sessionToken"  => $resp->sessionToken,
             "amount"        => $resp->ooAmount,
             "message"       => "Success"
-            ]
-        );
+        ]);
     }
 }